+++ pkg/enums/ctxkey/context_key.go
@@ -1,0 +1,5 @@
+package ctxkey
+
+//go:generate go-enum --values --names --flag --nocase --noprefix
+// ENUM(content-length)
+type ContextKey string
