--- pkg/enums/key/context_key.go
@@ -1,5 +1,0 @@
-package key
-
-//go:generate go-enum --values --names --flag --nocase
-// ENUM(content-length)
-type ContextKey string
